Handle Interrupted builds explicitly, add retry budget and Blocked propagation
- Replace the catch-all `_ => {}` in recorder that silently dropped
  Interrupted builds with explicit handling for each interruption kind
- Add `is_retryable()` to DrvBuildInterruptionKind: Timeout, OOM, and
  ProcessDeath get up to 2 attempts (via FailedRetry), while Cancelled
  and SchedulerDeath propagate Blocked to dependents immediately
- When retries are exhausted, interrupted builds transition to
  Completed(Failure) and propagate TransitiveFailure normally
- Handle UnsatisfiableRequirements explicitly instead of silently
  dropping it
- Extract forge notification logic into forge_notify.rs to keep
  request_handler.rs under the 500-line limit
