+++ backend/server/src/db/model/build_event.rs
@@ -181,6 +181,20 @@ pub enum DrvBuildInterruptionKind {
     SchedulerDeath,
 }
 
+impl DrvBuildInterruptionKind {
+    /// Whether this interruption kind is transient and the build should be
+    /// retried automatically. Non-retryable interruptions (Cancelled,
+    /// SchedulerDeath) propagate `Blocked` to dependents immediately.
+    pub fn is_retryable(&self) -> bool {
+        matches!(
+            self,
+            DrvBuildInterruptionKind::OutOfMemory
+                | DrvBuildInterruptionKind::Timeout
+                | DrvBuildInterruptionKind::ProcessDeath
+        )
+    }
+}
+
 mod state {
     use octocrab::params::checks::{
         CheckRunConclusion as GHConclusion, CheckRunStatus as GHStatus,
