+++ backend/server/src/scheduler/recorder/request_handler.rs
@@ -5,10 +5,7 @@ use std::sync::Arc;
 use tracing::{debug, warn};
 
 use super::{RecorderTask, RecorderWorker};
-use crate::channels::types::ChannelTask;
-use crate::config::ChannelForge;
 use crate::db::model::build_event;
-use crate::github::GitHubTask;
 use crate::scheduler::ingress::IngressTask;
 
 impl RecorderWorker {
@@ -193,215 +190,116 @@ impl RecorderWorker {
                     },
                 }
             },
-            _ => {},
-        }
+            DBS::Interrupted(ref kind) => {
+                debug!(
+                    "Attempting to record interrupted build of {} ({:?})",
+                    build_id.derivation.store_path(),
+                    kind
+                );
 
-        if let Some(github_sender) = &self.github_sender {
-            // Check if we need to create check_runs for failures
-            // Only create check_runs if the build failed and no check_run exists yet
-            if task.result.is_failure() {
-                let existing_check_runs = self.db_service.check_runs_for_drv_path(drv).await?;
+                let current_drv = self
+                    .db_service
+                    .get_drv(drv)
+                    .await?
+                    .ok_or_else(|| anyhow::anyhow!("Drv not found: {}", drv.store_path()))?;
+                let old_state = current_drv.build_state.clone();
 
-                if existing_check_runs.is_empty() {
-                    // No check_run exists, we need to create one
-                    // Get job info to know which jobsets this drv belongs to
-                    let job_infos = self.db_service.get_job_info_for_drv(drv).await?;
-
-                    for job_info in job_infos {
-                        let create_task = GitHubTask::CreateFailureCheckRun {
-                            drv_id: Arc::clone(drv),
-                            jobset_id: job_info.jobset_id,
-                            job_attr_name: job_info.name.clone(),
-                            difference: job_info.difference,
-                        };
-                        if let Err(e) = github_sender.send(create_task).await {
-                            warn!(
-                                "Failed to send CreateFailureCheckRun for {}: {:?}",
-                                drv.store_path(),
-                                e
+                if kind.is_retryable() {
+                    // Retryable interruption (Timeout, OOM, ProcessDeath).
+                    // Use the same two-attempt budget as Completed(Failure):
+                    //   Building -> first interrupt  -> FailedRetry (re-queue)
+                    //   FailedRetry -> second interrupt -> Completed(Failure) + propagate
+                    match current_drv.build_state {
+                        DBS::FailedRetry => {
+                            // Second attempt also interrupted — treat as permanent failure
+                            debug!(
+                                "Second interruption ({:?}) for {}, marking as permanent failure",
+                                kind,
+                                drv.store_path()
                             );
-                        }
-                    }
-                }
-            }
-
-            // Only send UpdateBuildStatus if the drv has check_runs.
-            // Dependency drvs (from the BFS cascade) rarely have
-            // check_runs — skipping them avoids flooding the GitHub
-            // service channel with no-op tasks.
-            let has_check_runs = !self
-                .db_service
-                .check_runs_for_drv_path(drv)
-                .await?
-                .is_empty();
-            if has_check_runs {
-                let github_task = GitHubTask::UpdateBuildStatus {
-                    drv_id: Arc::clone(drv),
-                    status: task.result.clone(),
-                };
-                if let Err(e) = github_sender.send(github_task).await {
-                    warn!(
-                        "Failed to send GitHub update for {}: {:?}",
-                        drv.store_path(),
-                        e
-                    );
-                }
-            }
-
-            // When a drv with check_runs completes, re-dispatch
-            // other queued drvs in the same jobset(s) for build
-            // re-evaluation. This handles the cascade when icu78
-            // completes and its dependents become buildable.
-            if has_check_runs && task.result == DBS::Completed(DBR::Success) {
-                for job_info in &job_infos {
-                    let queued_jobs = self
-                        .db_service
-                        .get_new_or_changed_jobs(job_info.jobset_id)
-                        .await?;
-                    for queued_job in queued_jobs {
-                        // Only re-dispatch drvs that are still queued
-                        if let Some(drv_entry) =
-                            self.db_service.get_drv(&queued_job.drv_path).await?
-                        {
-                            if matches!(drv_entry.build_state, DBS::Queued) {
-                                let drv_id = Arc::new(queued_job.drv_path);
-                                if let Err(e) = self
-                                    .ingress_sender
-                                    .try_send(IngressTask::EvalRequest(drv_id))
-                                {
-                                    warn!(
-                                        "ingress queue full, dropped EvalRequest re-dispatch: {}",
-                                        e
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Check if this drv completion concludes any jobsets
-            // Only check if we've reached a terminal state
-            if task.result.is_terminal() {
-                let job_infos = self.db_service.get_job_info_for_drv(drv).await?;
-
-                for job_info in job_infos {
-                    // Check if all jobs in this jobset are concluded
-                    if self
-                        .db_service
-                        .all_jobs_concluded(job_info.jobset_id)
-                        .await?
-                    {
-                        // Determine conclusion based on new/changed job failures
-                        let has_failures = self
-                            .db_service
-                            .jobset_has_new_or_changed_failures(job_info.jobset_id)
+                            self.update_and_broadcast(
+                                drv,
+                                &old_state,
+                                &DBS::Completed(DBR::Failure),
+                            )
                             .await?;
 
-                        let conclusion = if has_failures {
-                            octocrab::params::checks::CheckRunConclusion::Failure
-                        } else {
-                            octocrab::params::checks::CheckRunConclusion::Success
-                        };
-
-                        // Get jobset info to get the job name
-                        let jobset_info =
-                            self.db_service.get_jobset_info(job_info.jobset_id).await?;
-
-                        let complete_task = GitHubTask::CompleteCIEvalJob {
-                            ci_check_info: Arc::new(crate::github::CICheckInfo {
-                                commit: jobset_info.sha.clone(),
-                                base_commit: None,
-                                owner: jobset_info.owner.clone(),
-                                repo_name: jobset_info.repo_name.clone(),
-                            }),
-                            job_name: jobset_info.job.clone(),
-                            conclusion: conclusion.into(),
-                        };
-
-                        if let Err(e) = github_sender.send(complete_task).await {
-                            warn!(
-                                "Failed to send CompleteCIEvalJob for jobset {}: {:?}",
-                                job_info.jobset_id, e
+                            let blocked_drvs = self.propagate_graph_failure(drv).await?;
+                            if !blocked_drvs.is_empty() {
+                                self.db_service
+                                    .insert_transitive_failures(drv, &blocked_drvs)
+                                    .await?;
+                            }
+                        },
+                        _ => {
+                            // First interruption — transition to FailedRetry and re-queue
+                            debug!(
+                                "Retryable interruption ({:?}) for {}, transitioning to \
+                                 FailedRetry",
+                                kind,
+                                drv.store_path()
                             );
-                        }
+                            self.update_and_broadcast(drv, &old_state, &DBS::FailedRetry)
+                                .await?;
 
-                        // Notify ChannelService that this jobset has
-                        // concluded. The recorder cannot know which
-                        // release channels (if any) care about this
-                        // commit, so it always emits; ChannelService
-                        // filters by matching `(forge, owner, repo)`
-                        // against its registry and discards
-                        // jobset_complete events for which no
-                        // in-flight Evaluating row exists.
-                        //
-                        // Today the recorder only sees GitHub-backed
-                        // jobsets (the eval service is the sole
-                        // producer and is GitHub-only), so the forge
-                        // is hard-coded to GitHub. PR 4 will broaden
-                        // this when GitLab/Gitea start producing
-                        // jobsets.
-                        if let Some(channel_sender) = &self.channel_sender {
-                            let channel_task = ChannelTask::JobsetComplete {
-                                forge: ChannelForge::GitHub,
-                                owner: jobset_info.owner.clone(),
-                                repo: jobset_info.repo_name.clone(),
-                                sha: jobset_info.sha.clone(),
-                            };
-                            if let Err(e) = channel_sender.send(channel_task).await {
-                                warn!(
-                                    "Failed to send ChannelTask::JobsetComplete for jobset {}: \
-                                     {:?}",
-                                    job_info.jobset_id, e
-                                );
-                            }
-                        }
+                            let task = IngressTask::CheckBuildable(Arc::clone(drv));
+                            self.ingress_sender.send(task).await?;
+                        },
+                    }
+                } else {
+                    // Non-retryable interruption (Cancelled, SchedulerDeath).
+                    // Record the interrupted state and propagate Blocked to dependents.
+                    debug!(
+                        "Non-retryable interruption ({:?}) for {}, propagating Blocked to \
+                         dependents",
+                        kind,
+                        drv.store_path()
+                    );
+                    self.update_and_broadcast(drv, &old_state, &task.result)
+                        .await?;
 
-                        // Check if this is a PR that should be auto-merged
-                        if !has_failures {
-                            // Try to find a PR for this commit
-                            if let Ok(Some(pr)) = crate::db::github::get_pr_by_head_sha(
-                                &jobset_info.sha,
-                                &jobset_info.owner,
-                                &jobset_info.repo_name,
-                                &self.db_service.pool,
-                            )
-                            .await
-                            {
-                                // Fire CheckAutoMerge if UI auto-merge is on or a
-                                // comment-merge is pending. The handler re-validates
-                                // all gates (SHA-drift included), so over-triggering
-                                // is safe — drifted requests cancel cleanly.
-                                let has_pending_comment_merge = pr.comment_merge_sha.is_some();
-                                let is_open = pr.state == "open";
-                                if is_open && (pr.auto_merge_enabled || has_pending_comment_merge) {
-                                    debug!(
-                                        "PR #{} eligible for auto-merge check (auto_merge={}, \
-                                         comment_merge_pending={}), scheduling",
-                                        pr.pr_number,
-                                        pr.auto_merge_enabled,
-                                        has_pending_comment_merge
-                                    );
-
-                                    let auto_merge_task = GitHubTask::CheckAutoMerge {
-                                        owner: jobset_info.owner.clone(),
-                                        repo_name: jobset_info.repo_name.clone(),
-                                        pr_number: pr.pr_number,
-                                    };
-
-                                    if let Err(e) = github_sender.send(auto_merge_task).await {
-                                        warn!(
-                                            "Failed to send CheckAutoMerge for PR #{}: {:?}",
-                                            pr.pr_number, e
-                                        );
-                                    }
-                                }
-                            }
-                        }
+                    let blocked_drvs = self.propagate_graph_failure(drv).await?;
+                    if !blocked_drvs.is_empty() {
+                        self.db_service
+                            .insert_transitive_failures(drv, &blocked_drvs)
+                            .await?;
                     }
                 }
-            }
+            },
+            DBS::UnsatisfiableRequirements => {
+                debug!(
+                    "Recording unsatisfiable requirements for {}",
+                    build_id.derivation.store_path()
+                );
+
+                let old_state = self
+                    .db_service
+                    .get_drv(drv)
+                    .await?
+                    .map(|d| d.build_state)
+                    .unwrap_or(DBS::Queued);
+
+                self.update_and_broadcast(drv, &old_state, &task.result)
+                    .await?;
+
+                let blocked_drvs = self.propagate_graph_failure(drv).await?;
+                if !blocked_drvs.is_empty() {
+                    self.db_service
+                        .insert_transitive_failures(drv, &blocked_drvs)
+                        .await?;
+                }
+            },
+            _ => {
+                warn!(
+                    "Unexpected recorder task state {:?} for {}",
+                    task.result,
+                    drv.store_path()
+                );
+            },
         }
+
+        self.notify_forge_and_channels(drv, task, &job_infos)
+            .await?;
 
         // Broadcast job stats updates for all affected jobs
         for job_info in job_infos {
